Refactor: 단일플레이리스트 컴포넌트
